+++ examples/server/src/resolvers/Mutation.ts
@@ -1,5 +1,5 @@
-import { MyContext } from "../context";
 import { ResolverMap } from "../generated/schema";
+import { MyContext } from "../MyContext";
 
 export const mutationResolvers: ResolverMap<MyContext> = {
   Mutation: {