+++ examples/server/src/resolvers/index.ts
@@ -1,5 +1,5 @@
-import { MyContext } from "../context";
 import { ResolverMap } from "../generated/schema";
+import { MyContext } from "../MyContext";
 import { mutationResolvers } from "./Mutation";
 import { queryResolvers } from "./Query";
 import { scalarResolvers } from "./scalars";