+++ src/util/index.ts
@@ -4,6 +4,7 @@ export * from "./cleanOperations";
 export * from "./DeferredPromise";
 export * from "./getDirective";
 export * from "./isAsyncIterator";
+export * from "./listen";
 export * from "./LoggerLike";
 export * from "./parseDirective";
 export * from "./toError";