+++ src/util/toError.ts
@@ -1,3 +1,7 @@
-export function toError(err: unknown) {
-  return err instanceof Error ? err : new Error(`Unknown error: ${err}`);
+export function toError<TErrorExtensions = { status?: number }>(
+  err: unknown
+): Error & TErrorExtensions {
+  return (
+    err instanceof Error ? err : new Error(`Unknown error: ${err}`)
+  ) as Error & TErrorExtensions;
 }